Running final training check
Broke resume_training, curriculum, and train into their own files
to make things more modular. Now just need to run the training to
make sure that the model is able to learn.
