+++ indigo_league/training/__init__.py
@@ -1,0 +1,3 @@
+from indigo_league.training.resume_training import resume_training
+from indigo_league.training.train import curriculum
+from indigo_league.training.train import train
